+++ backend/.env.example
@@ -1,0 +1,12 @@
+BOT_TOKEN=
+DEBUG=False
+
+REDIS_PORT=6379
+REDIS_HOST=api-redis-db
+REDIS_DB=0
+
+POSTGRES_USER=postgres
+POSTGRES_PASSWORD=postgres
+POSTGRES_DB=api_database
+POSTGRES_HOST=api-postgres-db
+POSTGRES_PORT=5432
