+++ backend/admin/README.md
@@ -1,0 +1,9 @@
+# API системы администрирования бота
+
+## Документация
+
+- [FastAPI](https://fastapi.tiangolo.com/)
+- [Pydantic](https://docs.pydantic.dev/latest/)
+- [Pydantic Settings](https://docs.pydantic.dev/latest/concepts/pydantic_settings/)
+- [Redis](https://redis.readthedocs.io/en/stable/examples/asyncio_examples.html)
+- [SQLAlchemy](https://docs.sqlalchemy.org/en/20/)
