--- backend/bot/.env.example
@@ -1,6 +1,0 @@
-# Переменные окружения для настройки бота
-BOT_TOKEN=your_bot_token_here
-
-# Логирование
-# Возможные значения: DEBUG, INFO, WARNING, ERROR, CRITICAL
-DEBUG=False  # Для продакшн установить в False
