+++ backend/bot/README.md
@@ -1,0 +1,5 @@
+# Телеграм-бот
+
+## Документация
+
+- [aiogram](https://aiogram.dev/)
