--- backend/bot/handlers/__init__.py
@@ -1,6 +1,0 @@
-"""
-Модуль обработчиков сообщений.
-"""
-from . import start, common
-
-__all__ = ["start", "common"]
