--- backend/bot/main.py
@@ -1,61 +1,0 @@
-import asyncio
-import logging
-from pathlib import Path
-from aiogram import Bot, Dispatcher
-from aiogram.client.default import DefaultBotProperties
-from aiogram.enums import ParseMode
-
-from handlers import start, common
-from config import BOT_TOKEN, DEBUG
-
-
-# Настройка логирования
-Path("logs").mkdir(exist_ok=True)
-logging.basicConfig(
-    level=logging.DEBUG if DEBUG else logging.INFO,
-    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
-    handlers=[
-        logging.StreamHandler(),  # Вывод в консоль
-        logging.FileHandler("logs/bot.log", encoding='utf-8')
-    ]
-)
-logger = logging.getLogger(__name__)
-
-
-async def main() -> None:
-    """Основная функция для запуска бота."""
-    # Инициализация бота
-    bot = Bot(
-        token=BOT_TOKEN,
-        default=DefaultBotProperties(parse_mode=ParseMode.HTML)
-    )
-
-    # Инициализация диспетчера
-    dp = Dispatcher()
-
-    # Регистрация роутеров
-    dp.include_router(start.router)
-    dp.include_router(common.router)
-
-    # Удаляем вебхуки и запускаем polling
-    await bot.delete_webhook(drop_pending_updates=True)
-    logger.info("Бот запущен!")
-
-    try:
-        await dp.start_polling(bot)
-    except Exception as e:
-        logger.error(f"Ошибка при работе бота: {e}")
-    finally:
-        logger.info("Бот остановлен")
-        await bot.session.close()
-
-
-if __name__ == "__main__":
-    logger.info("Запуск приложения")
-
-    try:
-        asyncio.run(main())
-    except KeyboardInterrupt:
-        logger.info("Бот остановлен пользователем (Ctrl+C)")
-    except Exception as e:
-        logger.critical(f"Критическая ошибка: {e}")
