+++ backend/bot/requirements.txt
