+++ backend/bot/src/handlers/__init__.py
@@ -1,0 +1,6 @@
+"""
+Модуль обработчиков сообщений.
+"""
+from . import start, common
+
+__all__ = ["start", "common"]
