+++ backend/bot/src/utils/texts.py
@@ -1,0 +1,18 @@
+TEXTS: dict = {
+    'main_menu': "🏠 Главное меню",
+    'start': (
+        "👋 Привет! Я бот, который поможет тебе.\n\n"
+        "Выбери одну из опций ниже, чтобы начать."
+    ),
+    'about_button': "ℹ️ О нас",
+    'about_text': ("ℹ️ Здесь мы напишем информацию о боте\n\n"),
+    'help_button': "❓ Помощь",
+    'help_text': (
+        "📋 Здесь мы укажем доступные команды:\n\n"
+        "/start - Запустить бота\n"
+        "/help - Помощь\n"
+        "/about - О боте"
+    ),
+    'echo': "Пока все клавиши не работают, но я получил ваше сообщение!",
+    'unsupported': "❌ Извините, я не понимаю этот тип сообщений."
+}
