+++ backend/configs/site.conf
@@ -1,0 +1,30 @@
+server {
+    listen        80 default_server;
+    listen        [::]:80 default_server;
+    server_name   _;
+    server_tokens off;
+
+    root /data;
+
+    location @admin {
+        proxy_pass http://admin:8080;
+    }
+
+    location /api {
+        try_files $uri @admin;
+    }
+
+    location ~* \.(?:jpg|jpeg|gif|png|ico|css|js)$ {
+        log_not_found off;
+        expires 90d;
+    }
+
+    error_page  404 /404.html;
+
+    # redirect server error pages to the static page /50x.html
+    #
+    error_page   500 502 503 504  /50x.html;
+    location = /50x.html {
+        root   html;
+    }
+}
